+++ src/renderer/main.tsx
@@ -1,19 +1,20 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { exp } from './experimental'
-import { printProcessLogLevelInfo } from '../shared/logger'
+import initWasm from '@deltachat/message_parser_wasm'
 
 import App from './App'
-import { runtime } from './runtime'
-import init from '@deltachat/message_parser_wasm'
 import initSystemIntegration from './system-integration'
+import { exp } from './experimental'
+import { printProcessLogLevelInfo } from '../shared/logger'
+import { runtime } from './runtime'
 
 async function main() {
   exp.help //make sure experimental.ts is used
   try {
     runtime.initialize()
     printProcessLogLevelInfo()
-    await init('./message_parser_wasm_bg.wasm')
+
+    await initWasm('./message_parser_wasm_bg.wasm')
 
     initSystemIntegration()